collocation api factory moved to common files
