--- src/js/tiles/core/concordance/apiFactory.ts
@@ -1,37 +1,0 @@
-/*
- * Copyright 2019 Tomas Machalek <tomas.machalek@gmail.com>
- * Copyright 2019 Institute of the Czech National Corpus,
- *                Faculty of Arts, Charles University
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { SourceInfoComponent } from '../../../common/tile';
-import { init as fcs1Views } from '../../../views/clarin/fcs1/explain';
-import { ViewUtils, IActionDispatcher } from 'kombo';
-import { GlobalComponents } from '../../../views/global';
-import { CoreApiGroup } from '../../../common/api/coreGroups';
-
-
-export function createSourceInfoViewInstance(apiIdent:string, dispatcher:IActionDispatcher, viewUtils:ViewUtils<GlobalComponents>):SourceInfoComponent {
-
-	const views = fcs1Views(dispatcher, viewUtils);
-
-	switch (apiIdent) {
-		case CoreApiGroup.FCS_V1:
-			return views.ExplainView;
-		default:
-			return null;
-	}
-
- }
